# Speed up `Counter.Descending` by around 60%