## Summary

<!-- Describe your changes in detail here, if it closes an open issue, include "Closes #<issue>" -->
Removed the priority queue implementation for a simple sort and its much faster!

```plaintext
❯ benchstat before.txt after.txt
goos: darwin
goarch: amd64
pkg: github.com/FollowTheProcess/collections/counter
cpu: Intel(R) Core(TM) i5-8257U CPU @ 1.40GHz
             │ before.txt  │              after.txt              │
             │   sec/op    │   sec/op     vs base                │
Descending-8   796.2n ± 1%   339.3n ± 1%  -57.39% (p=0.000 n=25)

             │ before.txt │             after.txt              │
             │    B/op    │    B/op     vs base                │
Descending-8   440.0 ± 0%   200.0 ± 0%  -54.55% (p=0.000 n=25)

             │ before.txt │             after.txt              │
             │ allocs/op  │ allocs/op   vs base                │
Descending-8   8.000 ± 0%   4.000 ± 0%  -50.00% (p=0.000 n=25)
```